chore: challenge 3 completed
Created a new file to separate code
Created extra functions for:
list users
list account owners
validate birthday
add colors to terminal prompt for better UX
