+++ Challenge01_BankAccount.py
@@ -1,35 +1,22 @@
 import datetime as dt
+import CreateUserAndAccount as cua
 
 menu = """
 Choose an option: 
-[D]eposit 
-[W]ithdraw
-[B]alance
-[CA]create account
-[CU]create user
-[LA]ist accounts
-[LU]list users
-[Q]uit
+[\033[32mD\033[0m]eposit 
+[\033[32mW\033[0m]ithdraw
+[\033[32mB\033[0m]alance
+[\033[32mCA\033[0m]create account
+[\033[32mCU\033[0m]create user
+[\033[32mLA\033[0m]ist accounts
+[\033[32mLU\033[0m]list users
+[\033[31mQ\033[0m]uit
 """
 
 balance = 0
 counter = 0
 DAILY_TRANSACTION_LIMIT = 10
 transactions: dict[int, dict[str, float | dt.datetime]] = {}
-
-user = {}
-user_counter: int
-
-
-def create_user(name: str, birthday: str, person_ID: str, address: str):
-    pass
-
-
-def create_account(agency: str, account_number: str, user: dict[str, str]):
-    pass
-
-
-## NOTE - Transactional functions
 
 
 def transactions_record(option: str, amount: float, date: dt.datetime) -> bool:
@@ -110,37 +97,44 @@ def balance_handler(balance: float) -> float:
     return balance
 
 
-## NOTE - Main program
+# NOTE - Main program
 
-while True:
-    print("=" * 70)
-    option = input(f"{menu}\n-> ")
-    if option.lower() == "d":
-        balance = deposit_handler(balance)
 
-    elif option.lower() == "w":
-        balance = withdraw_handler(balance)
-
-    elif option.lower() == "b":
+def main():
+    global balance
+    while True:
         print("=" * 70)
-        print("Transactions:")
-        balance = balance_handler(balance)
+        option = input(f"{menu}\n-> ")
+        if option.lower() == "d":
+            balance = deposit_handler(balance)
 
-    elif option.lower() == "ca":
-        pass
+        elif option.lower() == "w":
+            balance = withdraw_handler(balance)
 
-    elif option.lower() == "cu":
-        pass
+        elif option.lower() == "b":
+            print("=" * 70)
+            print("Transactions:")
+            balance = balance_handler(balance)
 
-    elif option.lower() == "la":
-        pass
+        elif option.lower() == "ca":
+            cua.create_account()
 
-    elif option.lower() == "lu":
-        pass
+        elif option.lower() == "cu":
+            cua.create_user()
 
-    elif option.lower() == "q":
-        print("\33[96mService terminated. \nHave a nice day.\33[0m")
-        break
+        elif option.lower() == "la":
+            cua.list_accounts()
 
-    else:
-        print("\033[1;31;43m!!!Invalid Option!!!\033[0m")
+        elif option.lower() == "lu":
+            cua.list_users()
+
+        elif option.lower() == "q":
+            print("\33[96mService terminated. \nHave a nice day.\33[0m")
+            break
+
+        else:
+            print("\033[1;31;43m!!!Invalid Option!!!\033[0m")
+
+
+# NOTE - Calling main
+main()
